- Criada a estrutura de arquivos da entidade Member; - adicionado autorizacao para acesso a dados; - criado middleware checkProjectOwner; - validado as autorizacao no controller Project; - Relacionado as entidades Project e ProjectMember; - Fazendo verificacao de autorizacao de membros do projetos.

+++ app/Http/Controllers/ProjectController.php
@@ -2,26 +2,45 @@
 
 namespace CodeProject\Http\Controllers;
 
+use CodeProject\Entities\User;
 use CodeProject\Repositories\ProjectRepository;
 use CodeProject\Services\ProjectService;
 use Illuminate\Http\Request;
+use LucaDegasperi\OAuth2Server\Authorizer;
 
 
 class ProjectController extends Controller
 {
 
+
+    /**
+     * @var Authorizer
+     */
+    private $authorizer;
+    /**
+     * @var ProjectRepository
+     */
     private $repository;
+    /**
+     * @var ProjectService
+     */
     private $service;
 
     /**
      * ProjectController constructor.
      * @param ProjectRepository $repository
      * @param ProjectService $service
+     * @param Authorizer $authorizer
      */
-    public function __construct(ProjectRepository $repository, ProjectService $service)
+    public function __construct(
+        ProjectRepository $repository,
+        ProjectService $service,
+        Authorizer $authorizer
+        )
     {
         $this->repository = $repository;
         $this->service = $service;
+        $this->authorizer = $authorizer;
     }
 
     /**
@@ -29,24 +48,12 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        return $this->repository->all();
-    }
+        $userId = $this->authorizer->getResourceOwnerId();
+        $member = User::find($userId);
 
-    /**
-     * @return mixed
-     */
-    public function ownerClientAll()
-    {
-        return $this->service->ownerClientAll();
-    }
+        //return $member->projects;
 
-    /**
-     * @param $id
-     * @return mixed
-     */
-    public function ownerClientShow($id)    {
-
-        return $this->service->ownerClientShow($id);
+        return $this->repository->with(['owner', 'client'])->findWhere(['owner_id' => $this->authorizer->getResourceOwnerId()]);
     }
 
     /**
@@ -66,6 +73,9 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if($this->checkProjectOwner($id)==false) {
+            return ['erro' => 'Access Forbidden'];
+        }
         return $this->service->update($request->all(), $id);
     }
 
@@ -75,6 +85,9 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
+       if($this->checkProjectPermissions($id)==false) {
+           return ['erro' => 'Access Forbidden'];
+       }
         return $this->service->show($id);
     }
 
@@ -84,6 +97,31 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
+        if($this->checkProjectOwner($id)==false) {
+            return ['erro' => 'Access Forbidden'];
+        }
         return $this->service->delete($id);
+    }
+
+    private function checkProjectOwner($projectId)
+    {
+        $userId =  $this->authorizer->getResourceOwnerId();
+        return $this->repository->isOwner($projectId, $userId);
+    }
+
+    private function checkProjectMember($projectId)
+    {
+        $userId =  $this->authorizer->getResourceOwnerId();
+        return $this->repository->hasMember($projectId, $userId);
+    }
+
+    private function checkProjectPermissions($projectId)
+    {
+        if($this->checkProjectOwner($projectId) or $this->checkProjectMember($projectId))
+        {
+            return true;
+        }
+
+        return false;
     }
 }
